Added white gear, created abstract class for server parameters page
